feat(aloxide-js): enhance SDK read/write function and clean code (#19)
